+++ bundle/src/main/java/com/adobe/acs/commons/util/ClusterLeader.java
@@ -2,14 +2,14 @@
  * #%L
  * ACS AEM Commons Bundle
  * %%
- * Copyright (C) 2013 Adobe
+ * Copyright (C) 2021 Adobe
  * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,10 +17,17 @@
  * limitations under the License.
  * #L%
  */
-/**
- * Miscellaneous Utilities.
- */
-@Version("4.6.0")
-
 package com.adobe.acs.commons.util;
-import org.osgi.annotation.versioning.Version;
+
+import org.apache.sling.discovery.DiscoveryService;
+import org.apache.sling.discovery.InstanceDescription;
+
+/**
+ * Marker interface.
+ * Available as OSGi service only on instances being the cluster leader.
+ * @see InstanceDescription#isLeader
+ * @see DiscoveryService#getTopology
+ */
+public interface ClusterLeader {
+
+}