+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com)
 
 ### Fixed
 - #2496 - Upgrade shaded Guava dependency to 30.1
+- #2498 - Potential NPE in RunnableOnMaster
+- #2492 - NPE in JcrPackageReplicationStatusEventHandler
 - #2494 - Fixed issue with Versioned ClientLib incompatibility on 6.5.7
 
 ## 4.11.0 - 2020-12-11